rpi/requirements.txt
# Requirements for the Raspberry Pi companion (used by rpi/app.py)
# Install these into your project virtualenv (pip) on development machines, or prefer apt packages on Raspberry Pi for some items.

Flask>=2.0
flask-cors
python-dotenv
# gpiozero provides GPIO button support. On Raspberry Pi prefer the OS package(s): python3-gpiozero python3-rpi.gpio
gpiozero

# picamzero for Pi Camera Zero support
picamzero

# Geolocation support for IP-based location
geocoder

# Serial communication for USB buttons
pyserial
